+++ prog/gameLibs/publicInclude/render/daBfg/nodeHandle.h
@@ -20,7 +20,7 @@ namespace dabfg
 class NodeHandle
 {
   friend class NameSpace;
-  friend NodeHandle register_external_node(NodeNameId name_id, uint32_t generation);
+  friend NodeHandle register_external_node(NodeNameId name_id, uint16_t generation);
 
   NodeHandle(detail::NodeUid id) : uid{id} {}
